implemented available spaces dialog and other minor adjustments to spaces
git-svn-id: https://svn.duraspace.org/duracloud-priv@731 3c0c4179-e91d-42e7-a082-8e2792ce3676

+++ branches/duracloud-153/duradmin/src/main/webapp/javascript/spaces-manager.js
@@ -879,6 +879,9 @@ $(document).ready(function() {
 		
 		//attach delete button listener
 		$(".delete-space-button",detail).click(function(evt){
+			if(!dc.confirm("Are you sure you want to delete \n" + space.spaceId + "?")){
+				return;
+			}
 			
 			dc.store.DeleteSpace(space, {
 				begin: function(){
@@ -969,6 +972,9 @@ $(document).ready(function() {
 
 		//attach delete button listener
 		$(".delete-content-item-button",pane).click(function(evt){
+			if(!dc.confirm("Are you sure you want to delete \n" + contentItem.contentId + "?")){
+				return;
+			}
 			dc.store.DeleteContentItem(contentItem, {
 				begin: function(){
 					dc.busy( "Deleting content item...");
@@ -1040,12 +1046,14 @@ $(document).ready(function() {
 				spaceId, 
 				{
 					begin: function(){
-						dc.busy();
+						dc.busy("Loading space...");
 						$(contentItemListStatusId).html("Loading...").fadeIn("slow");
 					},
 					success: function(space){
+						if(space != undefined || space == null){
+							$(contentItemListStatusId).html("Error: space not found.").fadeIn("slow");
+						}
 						dc.done();
-
 						loadHandler(space);
 						$(contentItemListStatusId).fadeOut("fast");
 					},
